Moves extension methods to their own namespace
